+++ src/vtex/getProfile/getProfile.test.ts
@@ -1,0 +1,9 @@
+import getProfile from './getProfile';
+
+describe('getProfile', () => {
+  it('should be able to get current user profile', async () => {
+    const profile = await getProfile('econverse');
+
+    expect(profile).toBeDefined();
+  });
+});