+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [vtex] Added `getProductSuggestions` method.
 - [vtex] Added `getProductVariations` method.
 - [vtex] Added `getCollectionShelf` method.
+- [vtex] Added `getProfile` method.
 - [vtex] Added `MasterData.put` method.
 - [vtex] Added `MasterData.patch` method.